# src/main/resources/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost
    description: Generated server url
...
paths:
...
  "/notice/make":
    get:
      tags:
        - Notice-Controller
      summary: 공지사항 작성 폼
      description: 공지사항을 작성할 수 있는 폼을 반환합니다.
      operationId: memberForm_2
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
    post:
      tags:
        - Notice-Controller
      summary: 공지사항 작성 처리
      description: 공지사항 작성 폼의 데이터를 처리하고 공지사항을 저장합니다.
      operationId: memberForm
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NoticeDto"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/members/new":
    get:
      tags:
        - Member-Controller
      summary: 회원 가입 폼
      description: 회원 가입을 위한 폼을 반환합니다.
      operationId: memberForm_3
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
    post:
      tags:
        - Member-Controller
      summary: 회원 가입 처리
      description: 회원 가입 폼을 처리합니다.
      operationId: memberForm_1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MemberFormDto"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/admin/item/{itemId}":
    get:
      tags:
        - Item-Controller
      summary: 상품 상세 정보 조회
      description: 상품 ID를 기반으로 상품의 상세 정보를 조회합니다.
      operationId: itemDtl_1
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
    post:
      tags:
        - Item-Controller
      summary: 상품 정보 수정
      description: 상품의 정보를 수정합니다. 상품 이미지도 함께 업데이트할 수 있습니다.
      operationId: itemUpdate
      requestBody:
        content:
          application/json:
            schema:
              required:
                - itemImgFile
              type: object
              properties:
                itemFormDto:
                  "$ref": "#/components/schemas/ItemFormDto"
                itemImgFile:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/admin/item/new":
    get:
      tags:
        - Item-Controller
      summary: 상품 등록 폼
      description: 새 상품을 등록하기 위한 폼 페이지를 반환합니다.
      operationId: itemForm
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
    post:
      tags:
        - Item-Controller
      summary: 상품 등록
      description: 새 상품을 등록합니다. 상품 이미지도 함께 업로드할 수 있습니다.
      operationId: itemNew
      requestBody:
        content:
          application/json:
            schema:
              required:
                - itemImgFile
              type: object
              properties:
                itemFormDto:
                  "$ref": "#/components/schemas/ItemFormDto"
                itemImgFile:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/notice/notice/items/{page}":
    get:
      tags:
        - Notice-Controller
      summary: 공지사항 목록 조회
      description: 페이징 처리된 공지사항 목록을 조회합니다.
      operationId: noticeView_1
      parameters:
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/item/{itemId}":
    get:
      tags:
        - Item-Controller
      summary: 상품 상세 페이지 조회
      description: 고객이 상품 상세 페이지를 조회할 때, 최근 본 상품 목록에 추가합니다.
      operationId: itemDtl
      parameters:
        - name: member
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/MemberFormDto"
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: recentItems
          in: cookie
          required: false
          schema:
            type: string
            default: ''
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/admin/items/{page}":
    get:
      tags:
        - Item-Controller
      summary: 상품 관리 목록
      description: 상품 검색 조건에 맞는 상품 목록을 페이지네이션으로 관리합니다.
      operationId: itemManage
      parameters:
        - name: itemSearchDto
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/ItemSearchDto"
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/admin/items":
    get:
      tags:
        - Item-Controller
      summary: 상품 관리 목록
      description: 상품 검색 조건에 맞는 상품 목록을 페이지네이션으로 관리합니다.
      operationId: itemManage_1
      parameters:
        - name: itemSearchDto
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/ItemSearchDto"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
components:
  schemas:
...
    NoticeDto:
      required:
        - content
        - title
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
    MemberFormDto:
      required:
        - address
        - email
        - name
        - password
        - telNumber
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        address:
          type: string
        telNumber:
          type: string
        role:
          type: string
          enum:
            - USER
            - ADMIN
...
    ItemFormDto:
      required:
        - category
        - itemDetail
        - itemNm
        - price
        - stockNumber
      type: object
      properties:
        id:
          type: integer
          format: int64
        itemNm:
          type: string
        price:
          type: integer
          format: int32
        itemDetail:
          type: string
        stockNumber:
          type: integer
          format: int32
        itemSellStatus:
          type: string
          enum:
            - SELL
            - SOLD_OUT
        category:
          type: string
        view:
          type: integer
          format: int64
        itemImgDtoList:
          type: array
          items:
            "$ref": "#/components/schemas/ItemImgDto"
        itemImgIds:
          type: array
          items:
            type: integer
            format: int64
    ItemImgDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        imgName:
          type: string
        oriImgName:
          type: string
        imgUrl:
          type: string
        repImgYn:
          type: string
...
    ItemSearchDto:
      type: object
      properties:
        searchDateType:
          type: string
        searchSellStatus:
          type: string
          enum:
            - SELL
            - SOLD_OUT
        searchBy:
          type: string
        searchQuery:
          type: string